// src/types/formData.ts
export interface FormData {
  nome: string;
  setor: string;
  trabalhoEquipe: string;
  entregas: string;
  opinioesDivergentes: string;
  proatividade: string;
  posturaTrabalho: string;
  situacoesNovas: string;
  pressaoOpcoes: any[];
  pressaoOutros: string;
  comunicacao: string;
  pontualidade: string;
  metas: string;
  qualidade: string;
  adaptabilidade: string;
  iniciativa: string;
  colaboracao: string;
  flexibilidade: string;
  responsabilidade: string;
  atencaoDetalhes: string;
  resolucaoProblemas: string;
  lideranca: string;
  criatividade: string;
  organizacao: string;
  eticaProfissional: string;
  aprendizagem: string;
  relacionamentoInterpessoal: string;
  gestaoTempo: string;
  tomadaDecisao: string;
  visaoEstrategica: string;
  satisfacaoGeral: string;
  aspectosPositivos: string;
  aspectosMelhorar: string;
  sugestoesMelhorias: string;
  recursos: string;
  treinamentos: string;
  comunicacaoInterna: string;
  reconhecimento: string;
  equilibrioVida: string;
  crescimentoProfissional: string;
  autonomia: string;
  feedbackRegular: string;
  colaboracaoEquipes: string;
  inovacao: string;
  diversidadeInclusao: string;
  sustentabilidade: string;
  transparencia: string;
  apoioGestao: string;
  objetivosProfissionais: string;
  habilidadesDesenvolv: string;
  interesseTreinamento: string;
  carreira: string;
  contribuicaoEmpresa: string;
  melhoriaProcessos: string;
  feedbackGestao: string;
  expectativasFuturas: string;
  comentarios: string;
  experienciaExtra: string;
}

export interface Screen {
  component: React.ComponentType<any>;
  title: string;
}